refactor: Move components out of components folder
