+++ src/app/app-routing.module.ts
@@ -1,11 +1,11 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { PageNotFoundComponent } from './components/page-not-found/page-not-found.component';
+import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
 
 const routes: Routes = [
     {
         path: 'converter',
-        loadComponent: () => import('./components/converter/converter.component').then(m => m.ConverterComponent),
+        loadComponent: () => import('./converter/converter.component').then(m => m.ConverterComponent),
     },
     {
         path: '',
