+++ src/app/exchange-rate-card/exchange-rate-card.component.ts
