+++ src/app/exchange-rate-chart/exchange-rate-chart.component.scss
