+++ src/app/exchange-rate-chart/exchange-rate-chart.component.ts
@@ -2,7 +2,7 @@ import { Component, Input, OnChanges } from '@angular/core';
 import { ExchangeRate } from 'src/app/models';
 import { NgxEchartsModule } from 'ngx-echarts';
 import { Observable, bufferCount, scan, tap } from 'rxjs';
-import { SubscriptionSupervisorComponent } from 'src/app/components/subscription-supervisor/subscription-supervisor.component';
+import { SubscriptionSupervisorComponent } from 'src/app/subscription-supervisor/subscription-supervisor.component';
 import * as echarts from 'echarts';
 
 @Component({
